diffusion_rlhf/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "diffusion-rlhf"
version = "0.1.0"
description = "DSPO + SDXL fine-tuning with human feedback"
readme = "README.md"
authors = [{name = "ML Team"}]
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "diffusers>=0.24.0",
    "transformers>=4.35.0",
    "accelerate>=0.24.0",
    "peft>=0.7.0",
    "open-clip-torch>=2.20.0",
    "safetensors>=0.4.0",
    "optuna>=3.4.0",
    "pillow>=9.0.0",
    "numpy>=1.21.0",
    "pandas>=1.5.0",
    "tqdm>=4.64.0",
    "tensorboard>=2.14.0",
    "wandb>=0.15.0",
    "pyyaml>=6.0.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.6.0",
    "seaborn>=0.12.0",
    "jsonlines>=3.1.0",
    "datasets>=2.14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.0.0",
]

[tool.black]
line-length = 88
target-version = ["py311"]
include = '\.pyi?$'

[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = []

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*.py" = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=dspo",
    "--cov-report=term-missing",
]
